Split `sampling` into three modules
Closes #6141
