+++ pymc/__init__.py
@@ -68,6 +68,8 @@ from pymc.model_graph import model_to_graphviz, model_to_networkx
 from pymc.plots import *
 from pymc.printing import *
 from pymc.sampling import *
+from pymc.sampling_predictive import *
+from pymc.sampling_utils import *
 from pymc.smc import *
 from pymc.stats import *
 from pymc.step_methods import *
