+++ pymc/sampling_jax.py
@@ -7,7 +7,8 @@ from functools import partial
 from typing import Any, Callable, Dict, List, Optional, Sequence, Union
 
 from pymc.initial_point import StartDict
-from pymc.sampling import RandomSeed, _get_seeds_per_chain, _init_jitter
+from pymc.sampling import _init_jitter
+from pymc.sampling_utils import RandomSeed, _get_seeds_per_chain
 
 xla_flags = os.getenv("XLA_FLAGS", "")
 xla_flags = re.sub(r"--xla_force_host_platform_device_count=.+\s", "", xla_flags).split()
