+++ pymc/smc/kernels.py
@@ -35,7 +35,7 @@ from pymc.aesaraf import (
 from pymc.backends.ndarray import NDArray
 from pymc.blocking import DictToArrayBijection
 from pymc.model import Point, modelcontext
-from pymc.sampling import sample_prior_predictive
+from pymc.sampling_predictive import sample_prior_predictive
 from pymc.step_methods.metropolis import MultivariateNormalProposal
 from pymc.vartypes import discrete_types
